pyproject.toml
[tool.poetry]
name = "jasmin-manage"
version = "0.1.0"
description = "API for management of resources on JASMIN."
authors = ["Matt Pryor <matt.pryor@stfc.ac.uk>"]
maintainers = ["Matt Jones <matthew.j.jones@stfc.ac.uk>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/cedadev/jasmin-manage"

packages = [
    { include = "jasmin_manage_site"},
    { include = "jasmin_manage"},
]

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.1"
django = "~4.2.11"
markdown = "^3.3.4"
django-admin-list-filter-dropdown = "^1.0.3"
django-admin-rangefilter = "^0.12.1"
django-oauth-toolkit = "^2.2.0"
djangorestframework = "^3.12.4"
drf-spectacular = "^0.27.1"
drf-nested-routers = "^0.93.3"
jasmin-auth-django = {git = "https://github.com/cedadev/jasmin-auth-django.git", rev = "83ccea6ec052032cbdc0b52e856a31146e6e4d37"}
django-tsunami = {git = "https://github.com/cedadev/django-tsunami.git", rev = "54eaed8"}
django-tsunami-notify = {git = "https://github.com/cedadev/django-tsunami-notify.git", rev = "78e704c"}
django-settings-object = {git = "https://github.com/cedadev/django-settings-object.git", rev = "2b66c0fc5eae92972df5210b4bc43f7d95ad9ceb"}
django-debug-toolbar = "^4.3.0"

# Required for migrations for now.
django-markupfield = "^2.0.0"
django-concurrency = "^2.3"

[tool.poetry.group.test.dependencies]
tblib = "^3.0.0"
pytest = "^8.0.1"
coverage = "^7.4.3"

[tool.poetry.group.dev.dependencies]
tox = "^4.12.1"
black = "^24.1.1"
isort = "^5.12.0"
pytest = "^8.0.1"
pylint-django = "^2.5.5"
coverage = "^7.4.3"

[tool.black]
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"

[tool.pylint.main]
load-plugins = 'pylint_django'

[build-system]
requires = ["poetry-core", "setuptools>=59.6.0"]
build-backend = "poetry.core.masonry.api"
